Add AsnNull and AsnInteger decoding.

+++ trunk/asn1runtime/src/org/asn1gen/runtime/AsnInteger.scala
@@ -1,6 +1,6 @@
 package org.asn1gen.runtime
 
-case class AsnInteger(value: Int) extends AsnType {
+case class AsnInteger(value: Long) extends AsnType {
 }
 
 object AsnInteger extends AsnInteger(0) {
